layer dependencies fixed
