+++ packages/api/src/application/weather/interfaces/weather-provider.abstract.ts
@@ -1,4 +1,4 @@
-import { Weather } from '../weather.entity';
+import { Weather } from 'src/domain/weather/weather.entity';
 
 export abstract class WeatherProvider {
   abstract getCurrentWeather(city: string): Promise<Weather>;
