+++ packages/api/src/infrastructure/subscription/repositories/subscription.repository.ts
@@ -2,8 +2,8 @@ import { Injectable } from '@nestjs/common';
 import { InjectModel } from '@nestjs/mongoose';
 import { Model, RootFilterQuery } from 'mongoose';
 import { CreateSubscriptionDto } from 'src/application/subscriptions/dtos/create-subscription.dto';
+import { ServiceSubscriptionRepository } from 'src/application/subscriptions/interfaces/subscription-repository.abstract';
 import { HOUR } from 'src/common/utils/time-units';
-import { ServiceSubscriptionRepository } from 'src/domain/subscription/interfaces/subscription-repository.abstract';
 import { Subscription } from 'src/domain/subscription/subscription.entity';
 import { SubscriptionEntityMapper } from '../mappers/subscription-entity.mapper';
 import { SubscriptionDb } from '../schemas/subscription.schema';
@@ -11,7 +11,7 @@ import { SubscriptionDb } from '../schemas/subscription.schema';
 @Injectable()
 export class SubscriptionRepository implements ServiceSubscriptionRepository {
   constructor(@InjectModel(SubscriptionDb.name) private readonly subscriptionModel: Model<SubscriptionDb>) {}
-  async find(options: RootFilterQuery<SubscriptionDb>): Promise<Subscription[]> {
+  async find(options: RootFilterQuery<Subscription>): Promise<Subscription[]> {
     const found = await this.subscriptionModel.find(options);
     return found.map(SubscriptionEntityMapper.toEntity);
   }
@@ -26,7 +26,7 @@ export class SubscriptionRepository implements ServiceSubscriptionRepository {
     return SubscriptionEntityMapper.toEntity(savedSubscription);
   }
 
-  async updateById(id: string, updateDto: Partial<SubscriptionDb>): Promise<Subscription | null> {
+  async updateById(id: string, updateDto: Partial<Subscription>): Promise<Subscription | null> {
     const updated = await this.subscriptionModel.findByIdAndUpdate(id, updateDto).exec();
     if (!updated) return null;
     return SubscriptionEntityMapper.toEntity(updated);
