+++ packages/api/src/presentation/weather.controller.ts
@@ -1,13 +1,12 @@
 import { Controller, Get, Query } from '@nestjs/common';
-import { WeatherProvider } from 'src/domain/weather/interfaces/weather-provider.abstract';
-import { Weather } from 'src/domain/weather/weather.entity';
+import { WeatherProvider } from 'src/application/weather/interfaces/weather-provider.abstract';
 
 @Controller('weather')
 export class WeatherController {
   constructor(private readonly weatherProvider: WeatherProvider) {}
 
   @Get()
-  async getCurrentWeather(@Query('city') city: string): Promise<Weather> {
+  async getCurrentWeather(@Query('city') city: string) {
     return this.weatherProvider.getCurrentWeather(city);
   }
 }
