+++ packages/api/test/integration/subscription.integration.spec.ts
@@ -7,13 +7,13 @@ import { Model, Types } from 'mongoose';
 import { CityTestModule } from 'src/application/city/city.module.test';
 import { NotificationSubjects } from 'src/application/notifications/constants/notification-subjects.enum';
 import { NotificationType } from 'src/application/notifications/constants/notification-type.enum';
+import { NotificationsServiceInterface } from 'src/application/notifications/interfaces/notifications-service.abstract';
 import { NotificationsTestModule } from 'src/application/notifications/notifications.module.test';
 import { CreateSubscriptionDto } from 'src/application/subscriptions/dtos/create-subscription.dto';
 import { SubscriptionTestModule } from 'src/application/subscriptions/subscriptions.module.test';
 import { ExternalApiException } from 'src/common/errors/external-api.error';
 import { DatabaseExceptionFilter } from 'src/common/filters/database-exception.filter';
 import { appTestConfig, databaseTestConfig } from 'src/config/test.config';
-import { NotificationsServiceInterface } from 'src/domain/notifications/notifications-service.abstract';
 import { Frequency } from 'src/domain/subscription/frequency.vo';
 import { OpenWeatherCityProvider } from 'src/infrastructure/city/providers/openweather.provider';
 import { WeatherApiCityProvider } from 'src/infrastructure/city/providers/weatherapi.provider';
